Add files via upload
The one problem I was having is that python does not have a built in switch. I using multiple if statements seemed inefficent but without a switch it appears to be the easiest way. Im still half way looking into it but as I dont play the game anymore its safe to assume that the code will stay half working.